+++ src/library/PostgREST/SchemaCache.hs
@@ -22,16 +22,18 @@ module PostgREST.SchemaCache
   ( SchemaCache(..)
   , TablesFuzzyIndex
   , querySchemaCache
+  , dumpSchemaCache
   , showSummary
   , decodeFuncs
   , QueryTimings(..)
   , queryTimingsWLabels
   ) where
 
-import           Data.Aeson ((.=))
+import           Data.Aeson ((.:), (.=))
 import qualified Data.Aeson as JSON
 
 import qualified Data.ByteString.Char8      as BS
+import qualified Data.ByteString.Lazy       as LBS
 import qualified Data.HashMap.Strict        as HM
 import qualified Data.HashMap.Strict.InsOrd as HMI
 import qualified Data.Set                   as S
@@ -93,6 +95,20 @@ instance JSON.ToJSON SchemaCache where
     , "dbMediaHandlers"   .= JSON.toJSON hdlers
     , "dbTimezones"       .= JSON.toJSON tzs
     ]
+
+instance JSON.FromJSON SchemaCache where
+  parseJSON = JSON.withObject "SchemaCache" $ \o -> do
+    tabs <- o .: "dbTables"
+    SchemaCache tabs
+      <$> o .: "dbRelationships"
+      <*> o .: "dbRoutines"
+      <*> o .: "dbRepresentations"
+      <*> o .: "dbMediaHandlers"
+      <*> o .: "dbTimezones"
+      <*> pure (tablesFuzzyIndex tabs)
+
+dumpSchemaCache :: SchemaCache -> LBS.ByteString
+dumpSchemaCache = JSON.encode
 
 showSummary :: SchemaCache -> Text
 showSummary (SchemaCache tbls rels routs reps mediaHdlrs tzs _) =
@@ -181,16 +197,21 @@ querySchemaCache conf@AppConfig{..} = do
     , dbMediaHandlers = HM.union mHdlers initialMediaHandlers -- the custom handlers will override the initial ones
     , dbTimezones = tzones
 
-    , dbTablesFuzzyIndex =
-        -- Only build fuzzy index for schemas with a reasonable number of tables
-        -- Fuzzy.FuzzySet is memory heavy we just don't use it for large schemas
-        Fuzzy.fromList <$> HM.filter ((< maxDbTablesForFuzzySearch) . length) (HM.fromListWith (<>) ((qiSchema &&& pure . qiName) <$> HM.keys tabsWViewsPks))
+    , dbTablesFuzzyIndex = tablesFuzzyIndex tabsWViewsPks
     }, qsTime)
   where
     schemas = toList configDbSchemas
     isLogDebug = configLogLevel == LogDebug
     sqlTimedStmt = sqlTimedStatement isLogDebug
     sleepCall = SQL.Statement "select pg_sleep($1 / 1000.0)" (param HE.int4) HD.noResult True
+
+tablesFuzzyIndex :: TablesMap -> TablesFuzzyIndex
+tablesFuzzyIndex tabs =
+  -- Only build fuzzy index for schemas with a reasonable number of tables
+  -- Fuzzy.FuzzySet is memory heavy we just don't use it for large schemas
+  Fuzzy.fromList <$>
+    HM.filter ((< maxDbTablesForFuzzySearch) . length)
+      (HM.fromListWith (<>) ((qiSchema &&& pure . qiName) <$> HM.keys tabs))
 
 -- | overrides detected relationships with the computed relationships and gets the RelationshipsMap
 getOverrideRelationshipsMap :: [Relationship] -> [Relationship] -> RelationshipsMap